29/6/26 changes added

+++ src/main/java/com/example/DeliveryNotificationModule/repository/VehicleDeliveryRepository.java
@@ -2,5 +2,5 @@ package com.example.DeliveryNotificationModule.repository;
 
 import com.example.DeliveryNotificationModule.entity.VehicleDelivery;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface VehicleDeliveryRepository extends JpaRepository<VehicleDelivery,Integer> {
+public interface VehicleDeliveryRepository extends JpaRepository<VehicleDelivery,Long> {
 }
